tools/python_pptx_converter/requirements.txt
pymupdf
python-pptx
winsdk
Pillow
